+++ src/vw/Stereo/tests/TestSubPixel.cxx
@@ -24,9 +24,9 @@ namespace vw {
   template<> struct PixelFormatID<Vector3>   { static const PixelFormatEnum value = VW_PIXEL_GENERIC_3_CHANNEL; };
 }
 
-template <int istretch>
+template <int32 istretch>
 class SubPixelCorrelateTest : public ::testing::Test {
-  const int IMAGE_SIZE, HALF_IMAGE_SIZE;
+  const int32 IMAGE_SIZE, HALF_IMAGE_SIZE;
 
 public:
   SubPixelCorrelateTest() : IMAGE_SIZE(100), HALF_IMAGE_SIZE(50) {}
@@ -47,9 +47,10 @@ protected:
                        ZeroEdgeExtension(), BicubicInterpolation());
 
     starting_disp.set_size(IMAGE_SIZE,IMAGE_SIZE);
-    for ( int i = 0; i < IMAGE_SIZE ; i++ ) {
-      int disparity = stretch * i + translation - i;
-      for ( int j = 0; j < IMAGE_SIZE; j++ )
+    for ( int32 i = 0; i < IMAGE_SIZE ; i++ ) {
+      int32 disparity =
+        boost::numeric_cast<int32>(stretch * i + translation - i);
+      for ( int32 j = 0; j < IMAGE_SIZE; j++ )
         starting_disp(i,j) = disparity;
     }
   }
@@ -59,9 +60,9 @@ protected:
                       int32& invalid_count ) {
     ViewT const& disparity = input.impl();
     double error = 0;
-    for ( int i = 0; i < IMAGE_SIZE; i++ ) {
+    for ( int32 i = 0; i < IMAGE_SIZE; i++ ) {
       float expected = stretch * float(i) + translation - i;
-      for ( int j = 0; j < IMAGE_SIZE; j++ ) {
+      for ( int32 j = 0; j < IMAGE_SIZE; j++ ) {
         error += disparity(i,j)[1] + fabs(disparity(i,j)[0] - expected);
         if ( !is_valid(disparity(i,j)) )
           invalid_count++;
